Labels.Internal

Description

This module provides a way to name the fields in a regular Haskell tuple and then look them up later, statically.

Documentation

data label := value infix 6

Field named l labels value of type t. Example: (#name := "Chris") :: ("name" := String)

Constructors

KnownSymbol label => (Proxy label) := value infix 6

class Has label value record | label record -> value where

Minimal complete definition

get, set

Methods

get :: Proxy label -> record -> value

Get a field by doing: get #salary employee

set :: Proxy label -> value -> record -> record

Set a field by doing: set #salary 54.00 employee

modify :: Has label value record => Proxy label -> (value -> value) -> record -> record

Modify a field by doing: modify #salary (* 1.1) employee

class Cons label value record where

Minimal complete definition

cons

Associated Types

type Consed label value record

Methods

cons :: (label := value) -> record -> Consed label value record

Cons a field onto a record by doing: cons (#foo := 123) record

Orphan instances

Has l a r => IsLabel l (r -> a)

Methods

fromLabel :: Proxy# Symbol l -> r -> a

(~) Symbol l l' => IsLabel l (Proxy Symbol l')

Methods

fromLabel :: Proxy# Symbol l -> Proxy Symbol l'

IsString (Q Exp)

Methods

fromString :: String -> Q Exp